TicVariables & TicSettings

+++ Drivers/uspi/include/uspi/usbtict834.h
@@ -45,7 +45,7 @@ void _USBTicT834Device (TUSBTicT834Device *pThis);
 boolean USBTicT834DeviceConfigure (TUSBDevice *pUSBDevice);
 boolean USBTicT834DeviceWriteReg (TUSBTicT834Device *pThis, u8 nCommand, u16 nValue, u32 nIndex, u16 nLength, u32 *nData);
 boolean USBTicT834DeviceReadReg (TUSBTicT834Device *pThis, u8 nCommand, u16 nValue, u32 nIndex, u16 nLength, u32 *nData);
-int USBTicT834DeviceControl (TUSBTicT834Device *pThis, u8 nReqType, u8 nCommand, u16 nValue, u32 nIndex, u16 nLength, u32 *nData);
+int USBTicT834DeviceControl (TUSBTicT834Device *pThis, u8 nReqType, u8 nCommand, u16 nValue, u32 nIndex, u16 nLength, u8 *nData);
 boolean USBTicT834DeviceReadString (TUSBTicT834Device *pThis, u8 nString, u8 *nData);
 char * USBTicT834DeviceGetSerialNumber(TUSBTicT834Device *pThis);
